Add validations

+++ PolyclinicSystem/Controllers/PolyclinicController.cs
@@ -1,7 +1,8 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using Models;
-using Models.ModelsDb;
+using Models.Validations;
 using Services;
 
 namespace PolyclinicSystem.Controllers
@@ -12,12 +13,14 @@ namespace PolyclinicSystem.Controllers
         private PolyclinicService _polyclinicService;
         private IMapper _mapper;
         private IWebHostEnvironment _webHostEvironment;
+        private PolyclinicValidator _polyclinicValidator;
 
-        public PolyclinicController(IMapper mapper, IWebHostEnvironment webHostEvironment)
+        public PolyclinicController(IMapper mapper, IWebHostEnvironment webHostEvironment, PolyclinicValidator polyclinicValidator)
         {
             _webHostEvironment = webHostEvironment;
             _mapper = mapper;
-            _polyclinicService = new PolyclinicService(_mapper, _webHostEvironment);            
+            _polyclinicService = new PolyclinicService(_mapper, _webHostEvironment);   
+            _polyclinicValidator = polyclinicValidator;
         }
 
         [HttpGet]
@@ -36,18 +39,33 @@ namespace PolyclinicSystem.Controllers
         [HttpPost]
         public async Task<IActionResult> AddPolyclinicAsync(Polyclinic polyclinic)
         {
+            if (_polyclinicValidator.Validate(polyclinic).Errors.Count != 0)
+            {
+                throw new ValidationException(_polyclinicValidator.Validate(polyclinic).Errors);
+            }
+
             return await _polyclinicService.AddPolyclinicAsync(polyclinic);
         }
 
         [HttpPut]
         public async Task<IActionResult> UpdatePolyclinicAsync(Polyclinic polyclinic)
         {
+            if (_polyclinicValidator.Validate(polyclinic).Errors.Count != 0)
+            {
+                throw new ValidationException(_polyclinicValidator.Validate(polyclinic).Errors);
+            }
+
             return await _polyclinicService.UpdatePolyclinicAsync(polyclinic);
         }
 
         [HttpDelete]
         public async Task<IActionResult> DeletePolyclinicAsync(Polyclinic polyclinic)
         {
+            if (_polyclinicValidator.Validate(polyclinic).Errors.Count != 0)
+            {
+                throw new ValidationException(_polyclinicValidator.Validate(polyclinic).Errors);
+            }
+
             return await _polyclinicService.DeletePolyclinicAsync(polyclinic);
         }
     }
